+++ packages/lumx-react/src/components/checkbox/Checkbox.tsx
@@ -5,7 +5,7 @@ import classNames from 'classnames';
 import { mdiCheck } from '@lumx/icons';
 
 import { Icon, InputHelper, InputLabel, Theme } from '@lumx/react';
-import { COMPONENT_PREFIX, CSS_PREFIX } from '@lumx/react/constants';
+import { COMPONENT_PREFIX } from '@lumx/react/constants';
 import { Comp, GenericProps, getRootClassName, handleBasicClasses } from '@lumx/react/utils';
 
 import uniqueId from 'lodash/uniqueId';
@@ -28,8 +28,6 @@ export interface CheckboxProps extends GenericProps {
     name?: string;
     /** Theme adapting the component to light or dark background. */
     theme?: Theme;
-    /** Whether custom colors are applied to this component or not. */
-    useCustomColors?: boolean;
     /** Native input value property. */
     value?: string;
     /** On change callback. */
@@ -73,7 +71,6 @@ export const Checkbox: Comp<CheckboxProps, HTMLDivElement> = forwardRef((props, 
         name,
         onChange,
         theme,
-        useCustomColors,
         value,
         ...forwardedProps
     } = props;
@@ -97,7 +94,6 @@ export const Checkbox: Comp<CheckboxProps, HTMLDivElement> = forwardRef((props, 
                     prefix: CLASSNAME,
                     theme,
                 }),
-                { [`${CSS_PREFIX}-custom-colors`]: useCustomColors },
             )}
         >
             <div className={`${CLASSNAME}__input-wrapper`}>